requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
scipy==1.11.4
tensorflow==2.15.0
keras==2.15.0
scikit-fuzzy==0.4.2
matplotlib==3.8.2
seaborn==0.13.1
joblib==1.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
supabase==2.3.2
postgrest==0.13.2
realtime==1.0.6
storage3==0.7.7
supafunc==0.3.3
